+++ ktrain/text/predictor.py
@@ -37,15 +37,17 @@ class TextPredictor(Predictor):
             texts = [texts]
         elif not isinstance(texts, np.ndarray) and not isinstance(texts, list):
             raise ValueError('data must be numpy.ndarray or list (of texts)')
-        loss = self.model.loss
-        treat_multilabel = False
-        if loss != 'categorical_crossentropy' and not return_proba:
-            return_proba=True
-            treat_multilabel = True
+        classification, multilabel = U.is_classifier(self.model)
+        if multilabel: return_proba = True
+        #treat_multilabel = False
+        #loss = self.model.loss
+        #if loss != 'categorical_crossentropy' and not return_proba:
+        #    return_proba=True
+        #    treat_multilabel = True
         texts = self.preproc.preprocess(texts)
         preds = self.model.predict(texts)
         result =  preds if return_proba else [self.c[np.argmax(pred)] for pred in preds] 
-        if treat_multilabel:
+        if multilabel:
             result =  [list(zip(self.c, r)) for r in result]
         if is_str: return result[0]
         else:      return result