+++ CHANGELOG.md
@@ -9,8 +9,10 @@ Most recent releases are shown at the top. Each release shows:
 ## 0.2.5 (TBD)
 
 ### New:
+- ```explain``` methods have been added to ```TextPredictor``` and ```ImagePredictor``` objects.
 - ```TextPredictor.predict_proba``` and ```ImagePredictor.predict_proba_*``` convenience
 methods have been added.
+- Added ```utils.is_classifier``` utility function
 
 ### Changed:
 - ```TextPredictor.predict``` method can now accept a single document as input instead of